requirements.txt
pybind11==2.13.6
setuptools==80.9.0
pygame-ce==2.5.6